+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [4.3.46] - 2023-04-21
+### Changed
+- Appointments model synked with backend [#3149](https://github.com/rokwire/illinois-app/issues/3149).
+- My Groups empty message change [#3066](https://github.com/rokwire/illinois-app/issues/3066).
+### Fixed
+- Fixed system widgets update in Home panel [#3111](https://github.com/rokwire/illinois-app/issues/3111).
+- Handling tap action in inbox notifications [#3159](https://github.com/rokwire/illinois-app/issues/3159).
+- Fixed buildings target location evaluation for maps directions [#3100](https://github.com/rokwire/illinois-app/issues/3100).
 
 ## [4.3.45] - 2023-04-20
 ### Added